+++ cmtj/utils/procedures.py
@@ -1,25 +1,22 @@
 import math
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import Any, Dict, List, Tuple
+from typing import Any, Callable, Dict, List, Tuple
 
 import numpy as np
 from scipy.fft import fft, fftfreq
 from tqdm import tqdm
 
 from cmtj import AxialDriver, Axis, CVector, Junction, NullDriver, ScalarDriver
-from cmtj.utils.filters import Filters
 
 from .resistance import calculate_resistance_series, compute_sd
 
 
 @dataclass
 class ResistanceParameters:
-    """A data holder for resistance parameters. Not all have to be filled"""
+    """A data holder for resistance parameters. Not all have to be filled in."""
     Rxx0: float = 0
     Rxy0: float = 0
-    Rp: float = 100
-    Rap: float = 200
     Rahe: float = 0
     Rsmr: float = 0
     Ramr: float = 0
@@ -29,6 +26,7 @@ class ResistanceParameters:
 
 def compute_spectrum_strip(input_m: np.ndarray, int_step: float,
                            max_frequency: float):
+    """Compute the spectrum of a given magnetization trajectory."""
     yf = np.abs(fft(input_m))
     freqs = fftfreq(len(yf), int_step)
     freqs = freqs[:len(freqs) // 2]
@@ -52,10 +50,16 @@ def PIMM_procedure(
     simulation_duration: float = 5e-9,
     wait_time: float = 0e-9,
     max_frequency: float = 80e9,
+    resistance_fn: Callable = calculate_resistance_series,
     disturbance: float = 1e-3,
-    full_output: bool = False
+    take_last_n: int = 100,
+    full_output: bool = False,
+    disable_tqdm: bool = False,
+    static_only: bool = False
 ) -> Tuple[np.ndarray, np.ndarray, Dict[str, Any]]:
-    """Procedure for computing Pulse Induced Microwave Magnetometry
+    """Procedure for computing Pulse Induced Microwave Magnetometry.
+    It computes both PIMM and Resistance (for instance AHE loops).
+    Set `static_only` to True to only compute the static resistance.
     :param junction: junction to be simulated.
     :param Hvecs: list of cartesian vectors. (use FieldScan.amplitude_scan or alike)
     :param int_step: integration step [s].
@@ -65,8 +69,13 @@ def PIMM_procedure(
     :param wait_time: time to wait before taking vector for the fft [s].
     :param Hoe_duration: duration of Hoe excitation in multiples of in step
     :param max_frequency: maximum frequency -- larger will be dropped [Hz].
+    :param resistance_fn: function to be used to compute the resistance
+        (either calculate_resistance_series or calculate_resistance_parallel).
     :param disturbance: disturbance to be applied to the magnetization (std of normal distribution).
+    :param take_last_n: number of last time steps to be taken for the compuation.
     :param full_output: if True, return the full trajectories and per layer spectra.
+    :param disable_tqdm: if True, disable tqdm progress bar.
+    :param static_only: if True, only compute the static resistance.
     :return: (spectrum, frequencies, other_data)
     other_data is a dictionary with the following keys:
     - 'H': Hext field [A/m]
@@ -100,10 +109,15 @@ def PIMM_procedure(
 
     # get layer strings
     layer_ids = junction.getLayerIds()
+    if len(layer_ids) != len(resistance_params):
+        raise ValueError(
+            "The number of layers in the junction must match the number of resistance parameters!"
+        )
     output = defaultdict(list)
     normalising_factor = np.sum(
         [layer.thickness * layer.Ms for layer in junction.layers])
-    for H in tqdm(Hvecs, desc="Computing PIMM"):
+    freqs = None  # in case of static_only
+    for H in tqdm(Hvecs, desc="Computing PIMM", disable=disable_tqdm):
         junction.clearLog()
         junction.setLayerExternalFieldDriver(
             "all",
@@ -129,32 +143,34 @@ def PIMM_procedure(
             ]) * layer.thickness * layer.Ms / normalising_factor
             for layer in junction.layers
         ])
-        m = m_traj[:, :, -100:]  # all layers, all x, y, z, last 100 steps
-        mixed = np.asarray([
-            np.asarray(log[f"{layer.id}_m{extraction_m_component}"])[indx] *
-            layer.thickness * layer.Ms / normalising_factor
-            for layer in junction.layers
-        ])
-        mixed = np.squeeze(mixed)
-        mixed_sum = mixed.sum(axis=0)
-        yf, freqs = compute_spectrum_strip(mixed_sum, int_step, max_frequency)
+        m = m_traj[:, :,
+                   -take_last_n:]  # all layers, all x, y, z, last 100 steps
+        Rx, Ry = resistance_fn([r.Rxx0 for r in resistance_params],
+                               [r.Rxy0 for r in resistance_params],
+                               [r.Ramr for r in resistance_params],
+                               [r.Rahe for r in resistance_params],
+                               [r.Rsmr for r in resistance_params],
+                               m,
+                               l=[r.l for r in resistance_params],
+                               w=[r.w for r in resistance_params])
+        if not static_only:
+            mixed = np.asarray([
+                np.asarray(log[f"{layer.id}_m{extraction_m_component}"])[indx]
+                * layer.thickness * layer.Ms / normalising_factor
+                for layer in junction.layers
+            ])
+            mixed_sum = mixed.sum(axis=0)
+            yf, freqs = compute_spectrum_strip(mixed_sum, int_step,
+                                               max_frequency)
 
-        spectrum.append(yf)
-        Rx, Ry = calculate_resistance_series(
-            [r.Rxx0
-             for r in resistance_params], [r.Rxy0 for r in resistance_params],
-            [r.Ramr
-             for r in resistance_params], [r.Rahe for r in resistance_params],
-            [r.Rsmr for r in resistance_params],
-            m,
-            l=[r.l for r in resistance_params],
-            w=[r.w for r in resistance_params])
+            spectrum.append(yf)
+
         # fill the output dict
         output['H'].append(H)
         output['Rx'].append(Rx)
         output['Ry'].append(Ry)
         output['m_avg'].append(m_traj[:, :, -1].sum(0))
-        if full_output:
+        if full_output and not static_only:
             output['m_traj'].append(m_traj)
             for li, layer_id in enumerate(layer_ids):
                 y, _ = compute_spectrum_strip(mixed[li], int_step,
@@ -171,12 +187,14 @@ def VSD_procedure(junction: Junction,
                   Hvecs: np.ndarray,
                   frequencies: np.ndarray,
                   int_step: float,
-                  resistance_params,
+                  resistance_params: List[ResistanceParameters] = [],
                   Hoe_direction: Axis = Axis.yaxis,
                   Hoe_excitation: float = 50,
                   simulation_duration: float = 30e-9,
                   disturbance: float = 1e-3,
-                  Rtype: str = 'Rz'):
+                  Rtype: str = 'Rz',
+                  resistance_fn: Callable = calculate_resistance_series,
+                  disable_tqdm: bool = False):
     """Procedure for computing Voltage-Spin Diode.
     We use the Oersted field sine exctitation to excite the system.
     :param junction: junction to be simulated.
@@ -188,11 +206,22 @@ def VSD_procedure(junction: Junction,
     :param Hoe_excitation: excitation amplitude of Hoe [A/m].
     :param simulation_duration: duration of simulation [s].
     :param disturbance: disturbance to be applied to the magnetization (std of normal distribution).
+    :param resistance_fn: function to be used to compute the resistance
+        (either calculate_resistance_series or calculate_resistance_parallel). Rz forces standard magnetores.
     :param Rtype: type of resistance to be used. (Rx Ry or Rz)
+    :param disable_tqdm: if True, disable tqdm progress bar.
     """
     layer_ids = junction.getLayerIds()
+    if Rtype == 'Rz' and len(layer_ids) > 2:
+        raise ValueError(
+            "Rz can only be used for 2 layer junctions. Use Rx or Ry instead.")
+    elif len(resistance_params) != len(layer_ids):
+        raise ValueError(
+            "The number of layers in the junction must match the number of resistance parameters!"
+        )
 
-    def simulate_VSD(H: np.ndarray, frequency: float, resistance_params):
+    def simulate_VSD(H: np.ndarray, frequency: float,
+                     resistance_params: ResistanceParameters):
 
         if Hoe_direction == Axis.zaxis:
             oedriver = AxialDriver(
@@ -230,20 +259,29 @@ def VSD_procedure(junction: Junction,
             log[f'{layer_ids[i]}_mz']
         ] for i in range(len(layer_ids))])
         if Rtype == 'Rz':
-            if isinstance(resistance_params,
-                          list) and len(resistance_params) > 1:
-                resistance_params = resistance_params[0]
-            R = log[f'R_{layer_ids[0]}_{layer_ids[1]}']
+            if len(layer_ids) > 1:
+                raise ValueError(
+                    "Rz can only be used for 2 layer junctions. One layer can be fictisious."
+                )
+            elif len(layer_ids) == 2:
+                R = log[f'R_{layer_ids[0]}_{layer_ids[1]}']
+            elif len(layer_ids) == 1:
+                R = log['Resistance']
+            else:
+                raise ValueError(
+                    "Resistance definition ambiguous!"
+                    "If you want to use Rz, you must provide"
+                    "a single resistance parameter set or set Rp Rap"
+                    " at junction creation.")
         else:
-            Rx, Ry = calculate_resistance_series(
-                [r.Rxx0 for r in resistance_params],
-                [r.Rxy0 for r in resistance_params],
-                [r.Ramr for r in resistance_params],
-                [r.Rahe for r in resistance_params],
-                [r.Rsmr for r in resistance_params],
-                m_traj,
-                l=[r.l for r in resistance_params],
-                w=[r.w for r in resistance_params])
+            Rx, Ry = resistance_fn([r.Rxx0 for r in resistance_params],
+                                   [r.Rxy0 for r in resistance_params],
+                                   [r.Ramr for r in resistance_params],
+                                   [r.Rahe for r in resistance_params],
+                                   [r.Rsmr for r in resistance_params],
+                                   m_traj,
+                                   l=[r.l for r in resistance_params],
+                                   w=[r.w for r in resistance_params])
             if Rtype == 'Rx':
                 R = Rx
             elif Rtype == 'Ry':
@@ -255,7 +293,8 @@ def VSD_procedure(junction: Junction,
         return vmix
 
     spectrum = np.zeros((len(Hvecs), len(frequencies)))
-    for hindx, H in enumerate(tqdm(Hvecs, "Computing VSD")):
+    for hindx, H in enumerate(
+            tqdm(Hvecs, "Computing VSD", disable=disable_tqdm)):
         for findx, f in enumerate(frequencies):
             spectrum[hindx, findx] = simulate_VSD(H, f, resistance_params)
     return spectrum